+++ packages/edit-navigation/src/components/menus-editor/index.js
@@ -16,7 +16,7 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import CreateMenuArea from './create-menu-area';
-import MenuEditor from '../menu-editor';
+import NavigationEditor from '../navigation-editor';
 
 export default function MenusEditor( { blockEditorSettings } ) {
 	const { menus, hasLoadedMenus } = useSelect( ( select ) => {
@@ -111,7 +111,7 @@ export default function MenusEditor( { blockEditorSettings } ) {
 				/>
 			) }
 			{ hasMenus && (
-				<MenuEditor
+				<NavigationEditor
 					menuId={ menuId }
 					blockEditorSettings={ blockEditorSettings }
 					onDeleteMenu={ ( deletedMenu ) => {
