+++ packages/edit-navigation/src/components/navigation-editor/block-editor-area.js
