+++ packages/edit-navigation/src/components/navigation-editor/navigation-structure-area.js
