+++ packages/edit-navigation/src/components/navigation-editor/shortcuts.js
