+++ packages/edit-navigation/src/components/navigation-editor/style.scss
