+++ packages/edit-navigation/src/style.scss
@@ -5,7 +5,7 @@
 }
 
 @import "./components/layout/style.scss";
-@import "./components/menu-editor/style.scss";
+@import "./components/navigation-editor/style.scss";
 @import "./components/menus-editor/style.scss";
 @import "./components/delete-menu-button/style.scss";
 @import "./components/notices/style.scss";
